move @annotate to py3compat
