+++ IPython/utils/py3compat.py
@@ -223,6 +223,21 @@ else:
                 filename = fname
             builtin_mod.execfile(filename, *where)
 
+
+def annotate(**kwargs):
+    """Python 3 compatible function annotation for Python 2."""
+    if not kwargs:
+        raise ValueError('annotations must be provided as keyword arguments')
+    def dec(f):
+        if hasattr(f, '__annotations__'):
+            for k, v in kwargs.items():
+                f.__annotations__[k] = v
+        else:
+            f.__annotations__ = kwargs
+        return f
+    return dec
+
+
 # Parts below taken from six:
 # Copyright (c) 2010-2013 Benjamin Peterson
 #
